refactor: :art: Shift code around
