+++ src/graphics/atlas.rs
@@ -17,17 +17,29 @@
 
 use super::autotile_ids::AUTOTILES;
 use super::Quad;
-use super::{
-    AUTOTILE_AMOUNT, AUTOTILE_FRAME_COLS, AUTOTILE_FRAME_WIDTH, AUTOTILE_ROW_HEIGHT,
-    HEIGHT_UNDER_AUTOTILES, MAX_SIZE, TILESET_WIDTH, TILE_SIZE, TOTAL_AUTOTILE_HEIGHT,
-};
+
+pub const MAX_SIZE: u32 = 8192; // Max texture size in one dimension
+pub const TILE_SIZE: u32 = 32; // Tiles are 32x32
+pub const TILESET_COLUMNS: u32 = 8; // Tilesets are 8 tiles across
+pub const TILESET_WIDTH: u32 = TILE_SIZE * TILESET_COLUMNS; // self explanatory
+
+pub const AUTOTILE_ID_AMOUNT: u32 = 48; // there are 48 tile ids per autotile
+pub const AUTOTILE_FRAME_COLS: u32 = TILESET_COLUMNS; // this is how many "columns" of autotiles there are per frame
+pub const AUTOTILE_AMOUNT: u32 = 7; // There are 7 autotiles per tileset
+
+pub const AUTOTILE_ROWS: u32 = AUTOTILE_ID_AMOUNT / AUTOTILE_FRAME_COLS; // split up the 48 tiles across each tileset row
+pub const AUTOTILE_ROW_HEIGHT: u32 = AUTOTILE_ROWS * TILE_SIZE; // This is how high one row of autotiles is
+pub const TOTAL_AUTOTILE_HEIGHT: u32 = AUTOTILE_ROW_HEIGHT * AUTOTILE_AMOUNT; // self explanatory
+pub const HEIGHT_UNDER_AUTOTILES: u32 = MAX_SIZE - TOTAL_AUTOTILE_HEIGHT; // this is the height under autotiles
+
+pub const AUTOTILE_FRAME_WIDTH: u32 = AUTOTILE_FRAME_COLS * TILE_SIZE; // This is per frame!
 
 use crate::prelude::*;
 
 use image::GenericImageView;
 use std::sync::Arc;
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct Atlas {
     pub atlas_texture: Arc<image_cache::WgpuTexture>,
     pub autotile_width: u32,
